# psk_bot/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSK Bot API",
...
    "version": "2.1.0",
    "contact": {
      "name": "PlanetSkool",
      "url": "https://planetskool.com"
    }
  },
  "servers": [
    {
      "url": "https://chatbot.planetskool.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:5173",
      "description": "Local development"
    }
  ],
...
  "paths": {
    "/api/v1/chat": {
      "post": {
        "tags": ["Robot API"],
        "summary": "Chat (full response)",
        "description": "Send a message and receive a **complete response** in a single JSON reply.\n\nThis is the primary endpoint for robot hardware — optimized for low latency.\n\n**How it works:**\n1. If `document_id` is provided, answers **strictly** from that document (no general knowledge fallback)\n2. If no relevant content is found in the document, the bot says so\n3. If no `document_id`, queries the LLM directly with general knowledge\n4. Simple greetings (\"hi\", \"hello\") get instant responses without LLM",
        "operationId": "robotChat",
        "security": [{"ApiKeyHeader": []}, {}],
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "description": "Client identifier for rate limiting (defaults to IP address)",
            "schema": {"type": "string", "example": "robot_01"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ChatRequest"},
              "examples": {
                "simple_question": {
                  "summary": "General question (direct LLM)",
                  "value": {
                    "message": "What is the capital of France?",
                    "session_id": "robot_01"
                  }
                },
                "document_question": {
                  "summary": "Question about a specific document",
                  "value": {
                    "message": "What is the leave policy?",
                    "document_id": "gentari_employee_handbook_e84d2400",
                    "session_id": "robot_01"
                  }
                },
                "greeting": {
                  "summary": "Greeting (instant response, no LLM)",
                  "value": {
                    "message": "Hello!",
                    "session_id": "robot_01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ChatResponse"},
                "examples": {
                  "document_answer": {
                    "summary": "Answer from document (RAG)",
                    "value": {
                      "response": "The annual leave policy allows employees to take up to 20 days of paid leave per year.",
                      "latency_ms": 2340,
                      "source": "document",
                      "session_id": "robot_01"
                    }
                  },
                  "llm_answer": {
                    "summary": "Answer from direct LLM",
                    "value": {
                      "response": "The capital of France is Paris.",
                      "latency_ms": 1800,
                      "source": "llm",
                      "session_id": "robot_01"
                    }
                  },
                  "greeting_answer": {
                    "summary": "Greeting (instant)",
                    "value": {
                      "response": "Hello! I'm PSK Bot. How can I assist you today?",
                      "latency_ms": 1,
                      "source": "greeting",
                      "session_id": "robot_01"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request — missing message or document not ready",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"},
                "example": {"error": "Message is required."}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"},
                "example": {"error": "Document not found."}
              }
            }
          },
          "429": {"$ref": "#/components/responses/RateLimited"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": ["Robot API"],
        "summary": "Chat (SSE streaming)",
        "description": "Send a message and receive the response as a **stream of tokens** via Server-Sent Events (SSE).\n\nTokens stream directly from Ollama for minimal latency.\n\n**SSE Event Format:**\n```\ndata: {\"token\": \"Hello\", \"type\": \"token\"}\ndata: {\"token\": \" there\", \"type\": \"token\"}\ndata: {\"type\": \"done\", \"latency_ms\": 2340, \"source\": \"document\", \"session_id\": \"robot_01\"}\n```\n\n**Event types:**\n- `token` — A chunk of the response text\n- `done` — Stream complete, includes metadata\n- `error` — An error occurred during generation",
        "operationId": "robotChatStream",
        "security": [{"ApiKeyHeader": []}, {}],
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "description": "Client identifier for rate limiting",
            "schema": {"type": "string", "example": "robot_01"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ChatRequest"},
              "examples": {
                "stream_question": {
                  "summary": "Stream a document question",
                  "value": {
                    "message": "Summarize the corporate profile",
                    "document_id": "fm_48736d23",
                    "session_id": "robot_01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of tokens",
            "content": {
              "text/event-stream": {
                "schema": {"type": "string"},
                "example": "data: {\"token\": \"The \", \"type\": \"token\"}\n\ndata: {\"token\": \"company \", \"type\": \"token\"}\n\ndata: {\"type\": \"done\", \"latency_ms\": 2340, \"source\": \"document\", \"session_id\": \"robot_01\"}\n\n"
              }
            }
          },
          "400": {"description": "Bad request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": ["Robot API"],
        "summary": "List available documents",
        "description": "Returns all uploaded documents and their processing status.\n\nUse this to discover valid `document_id` values for the chat endpoints.\n\n**Document statuses:**\n- `ready` — Indexed and queryable\n- `processing` — Being ingested (with OCR if needed)\n- `failed` — Failed to process\n\n**Supported file types:** `.pdf`, `.docx`, `.txt`, `.md`, `.png`, `.jpg`, `.jpeg`, `.tiff`, `.tif`, `.bmp`, `.webp`\n\nPDF and image files are processed with OCR to extract text from scanned pages, charts, and diagrams.",
        "operationId": "robotListDocuments",
        "security": [{"ApiKeyHeader": []}, {}],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/DocumentListResponse"},
                "example": {
                  "documents": [
                    {"id": "fm_48736d23", "name": "Corporate Profile.pdf", "status": "ready"},
                    {"id": "handbook_e84d2400", "name": "Employee Handbook.pdf", "status": "ready"}
                  ]
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
...
    "/api/stream": {
      "post": {
        "tags": ["Web UI API"],
        "summary": "Chat (SSE streaming)",
        "description": "Send a question and receive tokens via SSE, streamed directly from Ollama.\n\n**SSE format:** `data: <token_text>\\n\\n`. Final event: `data: [DONE]\\n\\n`.",
        "operationId": "webStream",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": {"type": "string", "example": "What is the leave policy?"},
                  "history": {"type": "array", "items": {"type": "object", "properties": {"user": {"type": "string"}, "bot": {"type": "string"}}}, "default": []}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "SSE stream", "content": {"text/event-stream": {"schema": {"type": "string"}, "example": "data: The \n\ndata: leave \n\ndata: policy...\n\ndata: [DONE]\n\n"}}},
          "400": {"description": "Missing question", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for robot endpoints. Set `PSK_API_KEY` in `.env` to enable. If not set, endpoints are open."
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key",
        "description": "Alternative: pass API key as query parameter"
      }
    },
    "schemas": {
      "ChatRequest": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {"type": "string", "description": "The user's message or question.", "example": "What is the leave policy?"},
          "document_id": {"type": "string", "description": "Optional. ID of a specific document to query. When provided, the bot answers **strictly** from that document — no general knowledge fallback.\n\nGet valid IDs from `GET /api/v1/documents`.", "example": "fm_48736d23", "nullable": true},
          "session_id": {"type": "string", "description": "Optional. Client session identifier. Defaults to client IP.", "example": "robot_01"},
          "history": {
            "type": "array",
            "description": "Optional. Previous conversation turns for follow-up context.",
            "items": {"type": "object", "properties": {"role": {"type": "string", "enum": ["user", "assistant"]}, "content": {"type": "string"}}, "required": ["role", "content"]},
            "example": [{"role": "user", "content": "What is the leave policy?"}, {"role": "assistant", "content": "The leave policy includes..."}]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {"type": "string", "description": "Human-readable error message"}
        },
        "required": ["error"]
      }
    },
...
  }
}